+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 ## 1.0.0 (pending)
 
 Initial version. Roughly, this is a spin-off of the "stool build" command from 
-Stool 6 https://github.com/mlhartme/stool
+Stool 6 https://github.com/mlhartme/stool, but
+
+* build arguments no longer added as labels